server/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic==1.10.12
python-dotenv
